% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catcher.R
\name{init_argh}
\alias{init_argh}
\alias{stop_argh}
\title{Initialize argh}
\usage{
init_argh(lang = c("en", "ko"))

stop_argh()
}
\arguments{
\item{lang}{Language of the error to parse. Default
is english. Now support "en", "ko".}
}
\value{
an initialized session with argh
}
\description{
Initialize argh
}
\examples{
init_argh()
}
